packages/training/requirements.txt
ultralytics>=8.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
joblib>=1.3.0
requests>=2.31.0
tqdm>=4.66.0
gdown>=5.1.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
mat73>=0.62
h5py>=3.9.0
